+++ src/main/java/com/macys/survey/dao/SurveyCustomersDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.macys.survey.model.SurveyCustomersModel;
 
-public interface SurveyCustomersDao extends JpaRepository<SurveyCustomersModel, Integer>{
+public interface SurveyCustomersDao extends JpaRepository<SurveyCustomersModel, Long>{
 
 }
